A lot of stuff

+++ src/TGC.HomeAutomation/TGC.HomeAutomation.API/Humidity/HumidityEntity.cs
@@ -5,7 +5,7 @@ namespace TGC.HomeAutomation.API.Humidity;
 [TableItem("Temperatures")]
 public class HumidityEntity : AzureTableItem
 {
-	public float Humidity { get; set; }
+	public double Humidity { get; set; }
 	public Guid SensorId { get; set; }
 	public DateTime Created { get; set; }
 }
