+++ src/TGC.HomeAutomation/TGC.HomeAutomation.API/Temperature/TemperatureEntity.cs
@@ -5,7 +5,7 @@ namespace TGC.HomeAutomation.API.Temperature;
 [TableItem("Temperatures")]
 public class TemperatureEntity : AzureTableItem
 {
-	public float Temperature { get; set; }
+	public double Temperature { get; set; }
 	public Guid SensorId { get; set; }
 	public DateTime Created { get; set; }
 }
